+++ packages/svelte/src/index.ts
@@ -33,6 +33,7 @@ export { default as Skeleton } from "./Skeleton.svelte";
 export { default as Dialog } from "./Dialog.svelte";
 export { default as Toast } from "./Toast.svelte";
 export { default as Table } from "./Table.svelte";
+export { default as Timeline } from "./Timeline.svelte";
 export { default as Progress } from "./Progress.svelte";
 export { default as Slider } from "./Slider.svelte";
 export { default as ColorPicker } from "./ColorPicker.svelte";
@@ -99,6 +100,7 @@ export {
 	XtyleDialog,
 	XtyleToast,
 	XtyleTable,
+	XtyleTimeline,
 	XtyleProgress,
 	XtyleSlider,
 	XtyleColorPicker,